rpicam/requirements.txt
# Core dependencies
numpy>=1.21.0
opencv-python>=4.5.0
PyYAML>=6.0

# Raspberry Pi specific
picamera2>=0.3.0

# Hailo NPU (will be installed separately on Pi)
# hailo-platform>=4.14.0

# Development and testing
pytest>=7.0.0 